% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nifti.R
\docType{methods}
\name{check_nifti-methods}
\alias{check_nifti-methods}
\alias{check_nifti}
\alias{check_nifti,nifti-method}
\alias{check_nifti,character-method}
\alias{check_nifti,factor-method}
\alias{check_nifti,list-method}
\alias{check_nifti,array-method}
\alias{check_nifti,anlz-method}
\alias{check_nifti,ANY-method}
\title{Check if nifti image or read in a nifti image}
\usage{
check_nifti(
  x,
  reorient = FALSE,
  allow.array = FALSE,
  fast = FALSE,
  need_header = TRUE,
  ...
)

\S4method{check_nifti}{nifti}(
  x,
  reorient = FALSE,
  allow.array = FALSE,
  fast = FALSE,
  need_header = TRUE,
  ...
)

\S4method{check_nifti}{character}(
  x,
  reorient = FALSE,
  allow.array = FALSE,
  fast = FALSE,
  need_header = TRUE,
  ...
)

\S4method{check_nifti}{factor}(
  x,
  reorient = FALSE,
  allow.array = FALSE,
  fast = FALSE,
  need_header = TRUE,
  ...
)

\S4method{check_nifti}{list}(
  x,
  reorient = FALSE,
  allow.array = FALSE,
  fast = FALSE,
  need_header = TRUE,
  ...
)

\S4method{check_nifti}{array}(
  x,
  reorient = FALSE,
  allow.array = FALSE,
  fast = FALSE,
  need_header = FALSE,
  ...
)

\S4method{check_nifti}{anlz}(
  x,
  reorient = FALSE,
  allow.array = FALSE,
  fast = FALSE,
  need_header = TRUE,
  ...
)

\S4method{check_nifti}{ANY}(
  x,
  reorient = FALSE,
  allow.array = FALSE,
  fast = FALSE,
  need_header = TRUE,
  ...
)
}
\arguments{
\item{x}{character path of image or 
an object of class nifti, or array}

\item{reorient}{(logical) passed to \code{\link{readnii}} 
if the image
is to be re-oriented}

\item{allow.array}{(logical) Are array types allowed (TRUE) or
should there be an error if the object is not character or class
nifti.}

\item{fast}{if \code{TRUE}, then \code{\link{fast_readnii}} will be used 
versus \code{\link{readnii}} if the files need to be read in.}

\item{need_header}{if \code{TRUE}, then an image type with header information
will be returned.  If not, then an array is fine.  Used really only in 
conjunction with \code{allow.array}}

\item{...}{additional arguments to pass to \code{\link{readnii}} if relevant}
}
\value{
nifti object or array if allow.array=TRUE and x is an array
}
\description{
Simple check to see if input is character or of 
class nifti
}
\examples{
x = nifti()
check_nifti(x)
}
\seealso{
\link{readnii}
}
\author{
John Muschelli \email{muschellij2@gmail.com}
}
