% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_array.R
\name{ensure_array}
\alias{ensure_array}
\title{Ensure an array output}
\usage{
ensure_array(img)
}
\arguments{
\item{img}{Image object (\code{\link{nifti}} or \code{niftiImage})}
}
\value{
Array of same dimensions as image object
}
\description{
Forces an array output for manipulation and overall conversion
}
