% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_readnii.R
\name{fast_readnii}
\alias{fast_readnii}
\title{Reading NIfTI images through RNifti}
\usage{
fast_readnii(fname, dtype = TRUE, drop_dim = TRUE)
}
\arguments{
\item{fname}{file name of the NIfTI file.}

\item{dtype}{Should \code{\link{datatyper}} be run after reading?}

\item{drop_dim}{Should \code{\link{drop_img_dim}} be run after reading?}
}
\value{
A \code{nifti} object
}
\description{
This function calls the \code{\link{readNifti}} function from the 
\code{RNifti} package, and then converts the image to a \code{nifti} object
}
\examples{
set.seed(5)
dims = rep(10, 4)
arr = array(rpois(prod(dims), lambda = 2), dim = dims)
nim = oro.nifti::nifti(arr)
tfile = tempfile(fileext = ".nii.gz")
write_nifti(nim, tfile)
rimg = fast_readnii(tfile)
}
