% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images2matrix.R
\name{images2matrix}
\alias{images2matrix}
\title{Transform set of images to matrix}
\usage{
images2matrix(imgs, mask = NULL)
}
\arguments{
\item{imgs}{Vector of files or list of images (niftiImage, array, or nifti)}

\item{mask}{Binary image to subset the voxels}
}
\value{
Matrix of V by p, where V is the product of the dimensions of one 
image or the number of voxels in the mask, and p is the number of images
}
\description{
Creates a matrix, where the voxels are on the rows 
and images are on the columns
}
