% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_ts_to_list.R
\name{img_ts_to_list}
\alias{img_ts_to_list}
\title{Image Time Series to list}
\usage{
img_ts_to_list(imgs, copy_nifti = TRUE, warn = TRUE)
}
\arguments{
\item{imgs}{object of class \code{\link{nifti}} with 4 dimensions, 
aka a 4D time series}

\item{copy_nifti}{Should \code{nifti} objects be returned (\code{TRUE}) or 
simply arrays (\code{FALSE}).  Should only be used for slight speed up when
array is adequate}

\item{warn}{Should a warning be printed if object is not class
\code{\link{nifti}}}
}
\value{
List of images
}
\description{
Turns a 4D time series image to a list of 3D images
}
\note{
If the object is not of class \code{\link{nifti}} or have
4 dimensions, then the object is returned
}
\examples{
set.seed(5)
dims = rep(10, 4)
arr = array(rpois(prod(dims), lambda = 2), dim = dims)
nim = oro.nifti::nifti(arr)
simg = img_ts_to_list(nim)
simg_arr = img_ts_to_list(arr)
back = img_list_to_ts(simg_arr)
slist = lapply(simg, function(x) array(x, dim(x)))
testthat::expect_equal(slist, simg_arr)
simg_arr = img_ts_to_matrix(arr)
simg_arr = img_ts_to_df(arr)
}
