% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskEmptyImageDimensions.R
\name{maskEmptyImageDimensions}
\alias{maskEmptyImageDimensions}
\alias{mask_empty_dim}
\title{Replace Empty Image Dimensions with Mask Values}
\usage{
maskEmptyImageDimensions(img, ..., reorient = FALSE, mask.value = 0)

mask_empty_dim(img, ..., reorient = FALSE, mask.value = 0)
}
\arguments{
\item{img}{nifti object}

\item{...}{Arguments to be passed to \code{\link{getEmptyImageDimensions}}.}

\item{reorient}{Should image be reoriented if a filename?}

\item{mask.value}{Value to replace voxels outside the mask.}
}
\value{
Object of class \code{nifti} with the values of the 
original image in the image and \code{mask.value} outside the mask
}
\description{
Replaces values from dropped dimensions with a \code{mask.va.ue}
}
\note{
\code{mask_empty_dim} is a shorthand for \code{maskEmptyImageDimensions}
with all the same arguments.
}
\seealso{
\code{\link{getEmptyImageDimensions}}
}
