% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.nifti.R
\name{mean.nifti}
\alias{mean.nifti}
\alias{mean.anlz}
\title{Mean of Values in an Image}
\usage{
\method{mean}{nifti}(x, ..., mask)

\method{mean}{anlz}(x, ..., mask)
}
\arguments{
\item{x}{Object of class \code{nifti}}

\item{...}{Arguments passed to \code{\link{mean.default}}}

\item{mask}{object to subset the image.  If missing, then all 
values of the image are plotted.}
}
\value{
Output of \code{\link{mean}}
}
\description{
Computes the mean of values
of an image with the option for a mask.
}
\examples{
img = nifti(array(rnorm(10^3), dim = rep(10, 3)))
mask = img > 0
mean(img, mask = mask)
}
