% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_img.R
\name{quantile_img}
\alias{quantile_img}
\title{Create Quantile Image}
\usage{
quantile_img(img, mask = NULL, ...)
}
\arguments{
\item{img}{Character vector, or object of class \code{nifti}}

\item{mask}{Mask to determine cumulative distribution function (cdf) from}

\item{...}{Additional arguments to pass to \code{\link{ecdf}}}
}
\value{
Object of class \code{nifti}
}
\description{
Creates output image of the quantiles that each voxel is in, 
after applying the mask
}
\examples{
set.seed(5)
dims = rep(10, 3)
arr = array(rnorm(prod(dims)), dim = dims)
nim = oro.nifti::nifti(arr)
qimg = quantile_img(nim)
qarr = quantile_img(arr)
testthat::expect_equal(qarr, array(qimg, dim = dim(qarr)))
qimg = quantile_img(nim, mask = nim > 0)

}
