% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_nifti.R
\name{random_nifti}
\alias{random_nifti}
\title{Create Random `nifti` object}
\usage{
random_nifti(dim, ...)
}
\arguments{
\item{dim}{dimensions for the `nifti` object}

\item{...}{arguments to send to \code{\link[oro.nifti]{nifti}}}
}
\value{
A `nifti` object
}
\description{
Create Random `nifti` object
}
\examples{
random_nifti(c(10, 10, 2))
random_nifti(c(10, 10))
}
