% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rpi.R
\name{read_rpi}
\alias{read_rpi}
\title{Read NIfTI file reoriented to RPI}
\usage{
read_rpi(file, ..., verbose = TRUE)
}
\arguments{
\item{file}{file name of the NIfTI file.}

\item{...}{Arguments to pass to \code{\link{readnii}}}

\item{verbose}{print diagnostics, passed to \code{\link{orient_rpi_file}}}
}
\description{
This function calls the \code{\link{readnii}} function after
calling \code{\link{orient_rpi_file}} to force RPI orientation.
}
\note{
`read_rpi` uses `RNifti` to ensure the reading orientation
}
