% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nifti_header.R
\docType{methods}
\name{check_nifti_header-methods}
\alias{check_nifti_header-methods}
\alias{check_nifti_header}
\alias{check_nifti_header,nifti-method}
\alias{check_nifti_header,character-method}
\alias{check_nifti_header,factor-method}
\alias{check_nifti_header,list-method}
\alias{check_nifti_header,array-method}
\alias{check_nifti_header,anlz-method}
\alias{check_nifti_header,ANY-method}
\title{Check if nifti image or read in a nifti header}
\usage{
check_nifti_header(x)

\S4method{check_nifti_header}{nifti}(x)

\S4method{check_nifti_header}{character}(x)

\S4method{check_nifti_header}{factor}(x)

\S4method{check_nifti_header}{list}(x)

\S4method{check_nifti_header}{array}(x)

\S4method{check_nifti_header}{anlz}(x)

\S4method{check_nifti_header}{ANY}(x)
}
\arguments{
\item{x}{character path of image or 
an object of class nifti, or array}
}
\value{
nifti object or character
}
\description{
Simple check to see if input is character or of 
class nifti and read in the header
}
\author{
John Muschelli \email{muschellij2@gmail.com}
}
