% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_outside_hull.R
\name{replace_outside_surface}
\alias{replace_outside_surface}
\title{Replace Values Outside Surface of image}
\usage{
replace_outside_surface(
  img,
  value = 0,
  threshold = 0,
  replace_value = NA,
  reorient = FALSE
)
}
\arguments{
\item{img}{nifti object or array}

\item{value}{Value to check against.  If zero, then 
\code{replace_outside_surface} will include any dimension that has 
fewer than \code{threshold} zeroes. May be a vector of values, matched with 
\code{\link{match}}}

\item{threshold}{Include dimension if fewer than \code{threshold} voxels
are in the slice}

\item{replace_value}{Value to replace those outside the surface.}

\item{reorient}{Should image be reoriented if a filename?
Passed to \code{\link{check_nifti}}}
}
\value{
Creates an array of binary values.  If \code{img} is a \code{nifti}
object, then a \code{nifti} is returned
}
\description{
Determines values outside the surface of an image and gives a 
mask back with those values set to a replacement.
}
\examples{
set.seed(5)
dims = rep(10, 3)
arr = array(0, dim = dims)

arr[ 3:5, 4:6, c(2, 6:8, 5)] = 1
nim = oro.nifti::nifti(arr)
out = replace_outside_surface(nim, replace_value = 0)
out_arr = replace_outside_surface(arr, replace_value = 0)
testthat::expect_equal(out_arr, array(out, dim = dim(out)))
}
